Read one block at a time, to grep multi-line data.
The volume name and WWN values are in a block of the data file
on separate lines.
Each volume has a separate block w/ these data.
These blocks are separated from other file content by white space,
so read one entire block at a time and grep it for the data.