Working on volname & hostname confusion.
Found that saved data is being overwritten when one hostname has
multiple volname's related to it.

There are problems in the code related to volname vs hostname.
I began to fix some of those w/ this commit, but there are still
functionality problems.

Clarify volname vs hostname issues in future changes, & use the right
variable in the right place.

Make this commit, then merge it into the work branch because it has the
working code to extract WWN field, by using a regex that expects
newlines embedded in the text.