Use volname for key element in the data structure.
Change from using hostname to volname in the data structure,
and the program produces correct o/p.
Change the main array for storing data from @hosts to @volumes.

Use volname as key for the hash that stores data for each volume, and
the problems seen before are solved: no overwritten data.