Match volname to core hostname pattern to link them.
Prior code used explicit data found in the Mappings Table to link
related volnames and hostnames.  That table shows Domain name
instead of hostname, and the two might be quite different, so this
solution can have problems.

The code in this commit assumes that volname is always based on a
pattern that includes the core hostname.  After writing and testing it,
I found two volnames that are nothing like the hostname to which they
belong.  So this solution does not cover all cases either.

Save this code, in case it might be useful if I have to change the prior
approach.