Read i/p file once instead of twice.
Save the volname as a global var, and use it when saving WWN,
which is found just two lines later in the i/p file.  This structure
of the i/p file allows the program to match these two values w/o reading
the i/p file twice, and w/o reading it by paragraphs (instead of
single lines), which was the previous approach.

Remove trailing comma+space at end of each o/p line.

Update the reference o/p file for the test.

Add a bug note to codenotes, & mark it fixed.