Move regex matching code from get_host_stats to StorageArrayLog.pm
First commit of the StorageArrayLog.pm module w/ perl code.
Some of it is working, but the final o/p is not yet correct.

Added header lines required for the module.

Copied regex code directly from get_host_stats and modified it to work
in this module, after being called from the driver.

This is a step in building a module that can be used by other
driver code & applications.